+++ pkg/vadmin/create_db_at.go
@@ -39,12 +39,6 @@ func (a Admintools) CreateDB(ctx context.Context, opts ...createdb.Option) (ctrl
 	return ctrl.Result{}, nil
 }
 
-// CreateDB will construct a new DB using the vcluster-ops library
-func (v VClusterOps) CreateDB(ctx context.Context, opts ...createdb.Option) (ctrl.Result, error) {
-	v.Log.Info("Starting vcluster CreateDB")
-	return ctrl.Result{}, fmt.Errorf("not implemented")
-}
-
 // genCreateDBCmd will generate the command line options for calling admintools -t create_db
 func (a Admintools) genCreateDBCmd(s *createdb.Parms) []string {
 	cmd := []string{
